[ChaosVD] Changing the file name generation to be a name based on a time stamp and let UE trace define the full path.
The old implementation tried to create a name by appending a prefix "ChaosVD" to a number. That number was determined by trying file names until hitting a combination that does not exists (that implementation was taken as is from day one from the old serialization code for the Solver Evolution).

It seems there is a rare case where that will return a name UE trace cannot use because it thinks it already exists. Looking at the UE trace code, it does several validations to the path we are not making. Also, other tools seems to just provide the file name and let UE trace handle build the full path.

Therefore I am changing the file name generation code to just provide a base filename (with ChaosVD- appended), and using a timestamp to make it unique, like UE trace does. With this change the UE Trace code will handle building the full path.

The CVD files will still be written to the Profiling folder

#jira UE-210540, UE-210539, UE-210545
#tests I tested that I can start a recording to file, and the file will be written to the profiling folder as expected, and that I can also auto load them in CVD when the recording stops
#rb Benn.Gallagher

[CL 32587476 by sergio gardeazabal in ue5-main branch]