nDisplay: Fixed 2110 Split Frustum Tearing
Currently, the incoming texture may have extra black columns on the right side due to required padding. This will fixed at the Rivermax level in the future, but this change is a workaround that trims them based on the destination texture in nDisplay. For the cases of resolutions that are different, which would not be the viewport-viewport case, a cvar was provided to control how many rightmost columns to trim.

#jira UE-210191
#rb Alejandro.Arango, Andrey.Yamashev, Ruslan.Idrisov, Vitalii.Boiko
#author peter.tarasenko

[CL 32617178 by alejandro arango in 5.4 branch]