Shaders: Add a define to all shaders of the entry point name (defined as itself)
This allows the following pattern when there are multiple shader entrypoints in a single file:

#ifdef MyEntryPointName

void MyEntryPointName(...)
{
}

#endif

This makes it possible to completely hide code and variables that are specific to one shader type from the others, while keeping the convenience of having multiple shader definitions in a single .usf file. This pattern is completely opt-in and therefore does not disrupt any previous idioms, though it does mean some code could be simplified going forward.

Use this pattern to fix warnings from the new DXC compiler in lumen and distance field shaders stemming from a THREADGROUP_SIZE being re-used on an unused entrypoint and creating invalid dimensions. This warnings only occurs when the shader minifier is enabled, but with this new approach, the redundant code is stripped out before the minifier runs, which leads to more savings overall.


#rb graham.wihlidal, Krzysztof.Narkowicz

[CL 32654032 by chris kulla in ue5-main branch]