Hash phrases in phrase phase instead of entry phase.
Given the following two contiguous phrases in the data file A and B;

|AAAAaaa|bbbBBBB|

"aaabbb" represents a single cache item that was persisted to disk in two
writes - a "partial". Prior to this change the hash was taken at "aaa" (start
of entry). Now it is taken at "bbb" - the start of the phrase.

The journal will get two entries; one for "aaa" and one for "bbb". The first
will have a key of zero - entries with a null key are no longer added into the
DiskCache map. The entry for "bbb" will have the full key. It will also have an
non-zero offset. As the first entry of a phrase should have a zero offset, when
a non-zero offset is found this indicates a bias to find the start of a
partial.

This also works for huge partials that span three or more phrases;
"|AAAaa|bbb|ccCCCC|". "aa" and "bb" will have a zero key. "cc" will have a
non-zero offset that takes shows where "aa" is.

When loading, if the oldest detect phrase starts with a partial, that partial
is simply skipped.

#rb johan.berg
#rnx

[CL 32659360 by martin ridgers in ue5-main branch]