Cameras: camera rig asset editor
This big change implements a graph editor for camera rig assets. There are actually two editors, one for the node hierarchy and one for the transitions.

A common chunk of code implements a generic-ish "object tree" graph editor, which is used by both those graph editors. Given a few pieces of configuration, this mini-library provides a graph editor where nodes are objects, and links are references between these objects inside the same package.

An IObjectTreeGraphObject interface is provided to integrate between the editor and model layers, such as saving node positions and comments.

#jira UE-204848
#jira UE-209628

[CL 32688128 by ludovic chabant in ue5-main branch]