Removed `bSimpleQuery` from TEDS' queries.
Originally the idea was to have simple queries that map directly to Mass and complex queries have additional processing around them. This has not been implemented beyond the initial placeholders, which are currently causing invalid assert to trigger. Earlier change lists already removed the advanced query operators and introduced the query conditions which will be used for this in the future. This change removes the `bSimpleQuery` boolean, but will otherwise keep the SimpleQuery for now.

#jira UE-194182
#rb aditya.ravichandran, logan.buchy

[CL 32701265 by ronald koppers in ue5-main branch]