Chaos - Convex-Mesh Motion-Aware Collision Detection (MACD) improvements
For a Convex moving from X to P, MACD now chooses a point along X->P to detect collisions between the convex and each triangle. This is a cheap form of CCD. The chosen point is the closest point on X->P to the line through the triangle centroid along its normal. We are still selecting the convex and triangle features to project based on the nearest features at X, so it can generate contacts even if P is on the insde of the triangle (which is where hte non-MACD path fails).

#rb vincent.robert

[CL 32711781 by chris caulfield in ue5-main branch]