Do not attempt to load IAS TOCs that are newer than the current branch.
UnrealPak queries existing tables prior to uploading on-demand content so that it can deduplicate. All branches use the same S3 bucket, and FN/Main has a newer format of .ondemandtoc table files. Older branches cannot load this newer format but the check that a table's version is support was missing. This would cause UnrealPak to assert.

#rb per.larsson
#rnx

#ushell-cherrypick of 32787600 by Martin.Ridgers

[CL 32790791 by martin ridgers in ue5-main branch]