Replacing FMassActor with a tag when used in TEDS.
FMassActor was the last Mass element directly exposed to TEDS. It was causing issues as it's not a true TEDS column, so various checks couldn't be added as they would break on this fragment. It also diverges from other uses of UObject-based objects so required additional code. This change does not remove potentially unneeded code, but does remove all uses of FMassActor and removes the additional libraries this required pulling in.

#jira UE-194182
#rb logan.buchy

[CL 32796786 by ronald koppers in ue5-main branch]