Chaos - better initial overlap depentration
Initial overlap is now per shape pair rather than per-contact. This better handles the case where we initial generate a contact with an under-estimate of the penetration depth (because we corrected an edge collision), but later get a face collision with the correct deeper depth. Prior to this we wouild pop out by the delta between the two leading to a large velocity.

#rb benn.gallagher, vincent.robert

[CL 32811691 by chris caulfield in ue5-main branch]