Comment out assert which validates that buffers are created with a size that's a multiple of the stride.
This check was introduced in CL 32498769 as an attempt to catch bad usage of create/lock/unlock in the higher level code (that CL also fixes the one case which was found at the time). However, it turns out there multiple places in the high level rendering code which re-purpose buffers for a new format while there are still views with the old format lying around, and then lock them with a size computed based on the new stride, triggering the assert when the old views are re-created. This problem is benign as long as the old views aren't used again. It's too late in the current release cycle to fix the high level code, so we'll disable the check until the correct fix can be applied in Main.

#jira UE-211785
[FYI] dmitriy.dyomin
#rnx

[CL 32858915 by mihnea balta in ue5-main branch]