Path Tracer: Initial volumetric Cloud support
Support VolumetricCloud actors when using the reference atmosphere mode in the path tracer. Two code paths are supported. The first, reference mode invokes the cloud material via callable shaders. Unfortunately there is a fairly substantial overhead to invoking callable shaders from such a deeply nested inner loop. The other option (enabled by default) is to get cloud properties through a specially constructed "cloud map" which is a 3d texture with angular density so that the resolution is highest near the viewer. The default size of this map is 512x512x64 which is 256Mb (uint4 format). All the main physical properties of the cloud (extinction,albedo,dual-lobed phase function) are packed into this uint4 format and can be driven by the shader.

To optimize ray marching, a secondary 2d structure is built which captures both min/max density and z values per cell. This structure is axis-aligned and traversed with a 2D DDA to efficiently skip empty regions and restrict cloud ray marching to areas which are known to contain clouds. The memory usage for this structure at the default resolution of 512x512 (float4 format) is just 4Mb. This structure is incrementally updated during rendering, so it progressively fills in more accurately as more samples are taken. While there is a slight bias (and potentially missing details) at very low sample counts, the renderer is guaranteed to eventually converge when more samples are taken.

Because this feature is still experimental, it is disabled by default, but can be enabled at runtime with the cvar:

r.PathTracing.EnableReferenceClouds 1

This only works when reference atmosphere is also enabled.

Quality can be tuned with the following cvars:

r.PathTracing.CloudRoughnessCutoff 1 (make the clouds visible for rays of all roughness values, this makes the clouds visible to self-scattering which means true-multiple scattering within the cloud is computed. This can be very expensive for large cloudscapes.)
r.PathTracing.CloudMapResolution (tune the voxel x/y resolution of the clouds - default 512)
r.PathTracing.CloudMapDepth (tune the voxel z resolution of the clouds - default 64)
r.PathTracing.CloudMapEnable 0 (disables the voxelized representation of the clouds and run the full shader)

Also, the "Tracing Start Max Distance" is the main parameter which controls how "far" the clouds can be seen. This directly impacts the effective resolution, so keeping it as low as possible is essential.

High density clouds can show some missing voxels if the maximum number of ray march steps is exceeded (the default has been raised to 768 to accomodate the default volumetric cloud shader)

#rb Patrick.Kelly, Sebastien.Hillaire
#jira UE-211949

[CL 32859955 by chris kulla in ue5-main branch]