Cloth Editor: don't store the animation instance in the Preview scene, just allow the SkeletMeshComponent to keep track of it.
This fixes a problem where the animation playback controls don't respond after reloading the SkeletalMesh asset in the preview scene. When a SkelMeshComponent is re-registered (e.g. after the asset is reloaded), it unconditionally destroys and re-creates its AnimScriptInstance, so we're guaranteed to be holding onto the wrong object in the editor if we are tracking it ourselves.

#rb Alex.McAdams, kriss.gossart

[CL 32974841 by tyson brochu in ue5-main branch]