Fix issue where menu search fields were visible when they shouldn't
The issue happened when an FMenuBuilder built a menu widget of N items while the "Menu Search Field Visibility Threshold" setting was at some value lower than N - in this case the menu is created with a visible search field. If the user later changes the setting to a value higher than N, and the widget is still in the UI and not rebuilt, then the search field will still be visible but disabled.

This is fixed to properly hide the search field when the number of menu items is lower than the "Menu Search Field Visibility Threshold" setting by making the Visibility attribute of the SBox containing the search field read the setting. If the setting indicates the search field should be hidden, the SBox is collapsed and the search field disappears.

#jira UE-212148
#rb brooke.hubert

[CL 32999327 by sebastian arleryd in ue5-main branch]