PR #11769: Fixing Position Driven Light Functions For Path Tracing And Ray Tracing
The original patch was not quite correct, but pointed in the right direction. The miss shaders were not setting up ResolvedView which is necessary for various MaterialTemplate fields to be populated properly. They were also using WorldRayOrigin() instead of TranslatedWorldRayOrigin() which is not quite the same in the multi-view case (split screen rendering).

#jira UE-212701
#rb aleksander.netzel

[CL 33119086 by alister-chowdhury in ue5-main branch]