Fix: Deadlock when losing connection to ODSC server #jira UE-208926
As suggested by [at]benjamin.rouveyrol, clean up pending requests when the COTF connection thread is about to die.

Because our TPromise can't be abandoned, we need to assign some value to the Promises in the PendingRequests before we can destroy them. I chose to assign an error status response, similar to the response from a failed SendRequest. Since we assign values to the Promises, any threads waiting on those Promises' Futures are awakened. I also removed the assert failure on error response in ODSCThread and changed it to a log message and server connection reset, so that a future ODSC request would attempt to reestablish a connection to the server.
#rb Josie.Yang

[CL 33205076 by paul deberry in ue5-main branch]