In SSourceControlChangelistsWidget::OnRefreshSourceControlWidgets handle the detection of a duplicate item gracefully instead of with a check.
This case can occur when a submit is done external to the editor ( e.g. from P4V ) and has also been reported to happen after submits from within the editor.  Now, when it is detected we request a changelists refresh in the next tick by setting bShouldRefresh to true.

This changelist also fixes another issue where files in the list view would get cleared when the selected changelist changed and then never get re-added to the list view.

#jira UE-211943
#rb brooke.hubert, zach.rammell

[CL 33213432 by ross smith2 in ue5-main branch]