Sequencer: Fix move binding contents so that the source child tracks are moved to the destination child binding.
In 5.3, the conversion just calls Move Binding Contents from the source to the destination. The destination binding doesn't have any existing tracks.

In 5.4, there were some changes which result in OnActorAddedToSequencer now being called. That results in default tracks and then when Move Binding Contents is called, the source child tracks aren't copied over the default tracks. This is the issue that this CL fixes.

In Main, Move Binding Contents is never called because the conversion now calls AddOrReplaceBinding, so it's not a problem.

This is a redo of 19871735 and also fixes UE-145266

#jira UE-211555
#rb david.bromberg

[CL 33234445 by max chen in ue5-main branch]