Add favorites drop target for place actors
For toolbars, we need the ability to provide buttons with dropzone (or othere) capabilities, and we need to be able to track the selected index for the toolbar in the overflow button. I added the selected index for the toolbar, and for buttons on the toolbar I added a delegate to toolbar buttons which can decorate the button as needed (in this case adding a droptarget). In this case, using the asset droptarget is the first step that will be covered in this review, the intention is to ultimately support more than just assets for this droptarget.

The reason that we need to track the last selected index is because the overflow button in the SClippingVerticalBox needs to apply a certain style based upon whether something in the overflow has been selected. To enable that there are changes in SClippingVerticalBox to track and use the last selected index to provide the appropriate styling.

#jira UE-212486
#rb sebastian.arleryd

[CL 33268118 by karen jirak in ue5-main branch]