SUniformWrapPanel has jitter\scrollbar issues
SUniformWrapPanel, when made too small, can lose its panel layout when used with a scrollbar, or, alternatively, can develop a heavy "jitter". The reason for this is that while laying the children out, there is an edge case that is missed in displaying children, where none of the best fit cases work. We need to provide the most reasonable default view for when this happens. I will add some examples.

#jira UE-212482
#rb Patrick.Boutot

[CL 33285123 by karen jirak in ue5-main branch]