Add -PerformSlateFiltering flag to perforce-clang-format-diff.ps1 to skip formatting of declarative Slate code
By default, this new filter is turned off. When enabled, this flag runs a filter that strips out any Slate-like blocks from the output of p4 diff before it's given to clang-format-diff.py. For the purposes of this filter, Slate-like code blocks start with a line containing only a [ and end in a line containing only a ] or ];

[CL 33295876 by sebastian arleryd in ue5-main branch]