Skein / ViewportStatus: Fixed overlays remaining active after disabling revision control.
The OnChanged() delegate is somewhat misleading. It doesn't just fire when the CVar is changed, it fires whenever the CVar is set - even if it's set to the same value. This means that previously, several 'ApplyOverlays' queries would be registered simultaneously as the CVar gets sets several times, which could also be a cause of hitches. It also meant that the overlays kept being applied, even after revision control got disabled, because not all queries got unregistered.

#jira UE-213176
#rb Manuel.Lang

[CL 33298679 by wouter burgers in ue5-main branch]