Fix crash in JunoControlChannelComponent when subscribing to PT comm channel and garbage collecting JunoClusterUnionActor at the same time.
Need to give memory ownership of PT comm channel node to a PT scoped object. SimCallbackObject is a candidate because its memory is managed in the way that is needed for this case, however this isn't its intended use.

This also adds a change to SimCallbackObject which makes the PreSimulate function truly optional. The option specified in the template parameter already is checked before adding the simcallbackobject to the list of objects to tick, but it was not possible to compile a simcallbackobject without implementing an override of the PreSimulate function.

#rb cedric.caillaud

[CL 33310694 by steven barnett in ue5-main branch]