LandscapePatch: Make it possible to set the height and weight patch source modes in the blueprint editor. This will not do automatic texture copies on the archetype or allocate UTexture objects on the archetype. Instead the texture creation is defferred to component instancing.
For weight patches, remove automatic patch creation when adding to array in detail panel, because this causes stomping of instances in the world when editing values on the archetype. Instead, use creation through EditInlineNew. This also prevents blueprint weight patches from causing reference issues during autosave, as it sets the flags correctly.

Also add RF_Transactional to all the patch subobjects so that they properly work with undo.

#rb jonathan.bard

[CL 33341627 by semion piskarev in ue5-main branch]