Make concurrent scanning of arrays safe
#rb phil.pizlo jared.cotton

For concurrent marking of arrays to be safe, the GC needs to see a consistent
view of ArrayType, Num, and the value buffer itself. The prior scheme of concurrent
marking of VArrayBase did not achieve this. This patch makes it so that we store a header
in the value buffer itself. This new data structure is called VBuffer.

VBuffer's header contains: NumValues, Capacity, and ArrayType. ArrayType and Capacity are
immutable. NumValues is mutable.

This achieves safety in the concurrent GC because the GC will always see a valid VBuffer. So
when transitioning to a new VBuffer, we always ensure stores to the ArrayType/NumValues/Capacity
happen before the store of a new VBuffer happens on the array itself. We ensure this with proper
store fences.

This patch also makes it safe to append values into an existing VBuffer while concurrently scanning
said VBuffer. To ensure this is safe, the mutator will store fence between stores of the new VValue
and increments of NumValues. This ensures that when the GC sees a NumValues it is guaranteed that
VValues at 0..NumValues are non-garbage.

[CL 33361782 by saam barati in ue5-main branch]