Keep track of when a key state is flushed on the Player input functions. This can happen during several situations:
1. During a level transition
2. You manually call it in C++
3. You change the "input mode" of your player controller
4. The viewport focus is lost (i.e. you alt+Tab out of a window or otherwise click out of the game viewport window)
... probably several more ways as well.


When any of these things happen, if the user keeps holding down the key that has been flushed, then the Slate Application will send an IE_Repeat event, since it is what the actual operating system is reporting (and is correct). This means that your key state would get a repeat event before a "pressed" event. This puts the FKeyState in an incorrect state, and results in the improper evaluation of a "pressed" and "released" event firing every single frame, instead of a "repeat" event.

This has the side effect that any "timed" triggers in Enhanced Input (hold and release, hold, etc) would fire several delegates each frame because their state is basically re-evaluating the whole trigger set again, but it is correct because the starting time passes the actuation threshold that is specified.

This behavior is actually the same as we saw with UE-195988, and is behind the same CVar "Input.bAutoReconcilePressedEventsOnFirstRepeat" if you need to toggle it off. This is on by default, and this CVar will likely be removed in a future unless there are reports of broken behavior with it, as it is the "correct" behavior for the input system here from the engine standpoint.

Phew, das a lot of words... if you read this far... you get a cookie?

#jira UE-203642
#jira UE-210200
#pf 6631451f882a9e6121af7f8b
#rb jodon.karlik

[CL 33398406 by ben hoffman in ue5-main branch]