VerseVM race and rush
All structured concurrency forms now give the EndTask instruction an additional output register and semaphore operand. Completed tasks write their results to the output register if it is uninitialized (supporting rush) and signal the semaphore. The parent task may wait for either a single signal (supporting race and rush) or several (supporting sync).

This functionality is part of the EndTask instruction because it must happen after child tasks are canceled, and because the signaled task must be resumed after any other awaiters (if it were possible to await these tasks).

#rb Markus.Breyer

[CL 33487220 by russell johnston in ue5-main branch]