Fixed issues with this CL: Cvar access in ShouldBuildQuadTreeInSceneProxy() now uses GetValueOnAnyThread() instead of GetValueOnGameThread() since the latter managed to trigger an ensure(). UWaterMeshComponent::CalcBounds() now properly initializes the FBox so it doesn't trigger a NaN check on the CDO where we don't have an actual bounding box we could use.
[Backout] - CL33506540
[FYI] bob.tellez
Original CL Desc
-----------------------------------------------------------------
[Backout] - CL33457826
[FYI] tim.doerries
Original CL Desc
-----------------------------------------------------------------
Moved water quadtree creation to WaterMeshSceneProxy using a new FWaterQuadTreeBuilder class which can be used to construct quadtrees at different positions using a certain set of fixed parameters. This is the first step towards managing per-view quadtrees in the scene proxy. There should be no difference in behavior with/without this CL; this is merely a change in architecture to enable future improvements. There's a new cvar to toggle between the old and now path: r.Water.WaterMesh.BuildQuadTreeInSceneProxy. In order for that cvar to take effect, the water mesh needs to be rebuild. This can easily be achieved with r.Water.WaterMesh.ForceRebuildMeshPerFrame 1.

#rb roey.borsteinas
#jira UE-212105
#rnx
#tests Local testing on PC in editor (QA-Waterworks, helios_terrain and creating new water bodies in a new empty level; with and without far distance material/mesh) as well as in a cooked instance of QAGame running QA-Waterworks

[CL 33522287 by tim doerries in ue5-main branch]