Mass' chunk sizes aren't used by its derived archetypes.
When a new archetype is created by mutating an existing one, the new archetype doesn't use the chunk size of the original archetype but instead retrieves the size from the settings. This causes unexpected behavior as the intention is the automatically created archetype mirrors the original. This fix makes sure the chunk size is properly inherited.

#jira UE-214306
#rb Yoan.StAmant
[RN] Chunk sizes set on archetypes in Mass are now respected by automatically generated derived archetypes.

[CL 33579324 by ronald koppers in ue5-main branch]