Motion input (gyro, accelerometer) with supported controllers and mobile platforms now use the same axes and units by default. #jira UE-138975
In the past, we've been inconsistent with motion-related input on controllers and mobile platforms -- they've used different coordinate spaces and different units. This means users (us included) have had to manually convert accelerometer and gyro input for each platform on a per-game basis.

In 2022 we added the *.UnifyMotionSpace CVars, which let us get motion input in Unreal's coordinate space (left-handed, z-up, etc) when set to 1, or a similar but right-handed space when set to 2. Setting to non-zero also unifies the units across platforms, giving angular velocity in radians per second and acceleration in g-units.

For a better out-of-the-box experience, this change makes all of the *.UnifyMotionSpace CVars set to 1 by default.

[FYI] ben.hoffman
#tests PIE

[CL 33622860 by jibb smart in ue5-main branch]