Fix for trace snapshots on platforms that use internal send buffers
On some platforms we use an internal send buffer to reduce network overheads. When a trace is active and a snapshot was triggerered the send buffer could be flushed while writing the snapshot, making the ongoing trace miss events. To fix this add a forced flush right before creating the snapshot handle.

#rb Martin.Ridgers
#jira UE-211694

[CL 33626357 by johan berg in ue5-main branch]