Add the concept of section menus to ToolMenu sections
A section menu is currently only rendered for toolbar sections. The section menu can be opened via a new combo button that is added as the left-most entry of toolbar sections. When that combo button is clicked, a menu opens containing all entries of the section.

Section menus are disabled by default, but can be enabled on a section by calling the new FToolMenuSection::SetShowSectionMenu(bool) API.

#jira UE-212288
#rb aditya.ravichandran

[CL 33626872 by sebastian arleryd in ue5-main branch]